Implemented TextStep, made changes to Dialogue object and TalkStep abstract class. Changed the packaging of the entities due to the renaming of the file. Implemented the basic structures for the step entities for the call of usecase interactor.

+++ backend/src/main/java/com/speakfluid/backend/entities/Dialogue.java
@@ -19,7 +19,7 @@ public class Dialogue {
     // Store the dialogue between the chatBot and the user
     private ArrayList<Speech> dialogue;
     private String stepSuggestion;
-    private int confidenceScore;
+    private double confidenceScore;
 
 
     public Dialogue(ArrayList<Speech> dialogue){
@@ -30,7 +30,7 @@ public class Dialogue {
         return this.dialogue;
     }
 
-    public void setConfidenceScore(int confidenceScore) {
+    public void setConfidenceScore(double confidenceScore) {
         this.confidenceScore = confidenceScore;
     }
